Fixing warnings for a future compiler version, when the lab updates fedora version.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@931 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ src/GUI/EditRobotDialog.cpp
@@ -159,7 +159,7 @@ SetUpDialog() {
 
   m_jointConnectionsLine2->setReadOnly(true);
   m_jointConnectionsLine1->setValidator(
-      new QIntValidator(0, numeric_limits<size_t>::max(), this));
+      new QIntValidator(0, numeric_limits<int>::max(), this));
 
 
   jointAttributes->addWidget(new QLabel("<b>Limits:<\b>", this), 9, 1, 1, 4);
